db_pyqt_lesson_7_file_cleaning
